+++ CHANGELOG.rst
@@ -1,10 +1,12 @@
+2020/01/30 Version 7.0.0
+------------------------
 - Added a ``TrioEventEmitter`` class for intended use with trio
 - ``AsyncIOEventEmitter`` now correctly handles cancellations
 - Add a new experimental ``pyee.uplift`` API for adding new functionality to
   existing event emitters
 
 2019/04/11 Version 6.0.0
------------------------
+------------------------
 - Added a ``BaseEventEmitter`` class which is entirely synchronous and
   intended for simple use and for subclassing
 - Added an ``AsyncIOEventEmitter`` class for intended use with asyncio